requirements.txt
streamlit
python-dotenv
google.generativeai
streamlit-chat
streamlit_chat 
groq

